engine/common: make topo_sort more general so we can use it for models
